api/requirements.txt
art==6.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-oauth2-tool==0.0.3
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.2
oauth2client==4.1.3
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pymongo==4.12.1
pyparsing==3.2.3
requests==2.32.3
rsa==4.9.1
six==1.17.0
uritemplate==4.1.1
urllib3==2.4.0
waitress==3.0.2
Werkzeug==3.1.3
